requirements.txt
altgraph==0.17.4
click==8.1.7
colorama==0.4.6
install==1.3.5
keyboard==0.13.5
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.2
pyqt6-tools==6.4.2.3.3
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
matplotlib==3.8